# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Landmark {
  pageId: number;
  title: string;
  lat: number;
  lng: number;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  extract?: string | null;
  wikipediaUrl: string;
}

export interface LandmarkDetail {
  pageId: number;
  title: string;
  lat: number;
  lng: number;
  /** @nullable */
  extract?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  wikipediaUrl: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetLandmarksParams = {
/**
 * Northern latitude of the bounding box
 */
north: number;
/**
 * Southern latitude of the bounding box
 */
south: number;
/**
 * Eastern longitude of the bounding box
 */
east: number;
/**
 * Western longitude of the bounding box
 */
west: number;
};