# lib/api-zod/src/generated/types/landmark.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Landmark {
  pageId: number;
  title: string;
  lat: number;
  lng: number;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  extract?: string | null;
  wikipediaUrl: string;
}