# lib/api-zod/src/generated/types/landmarkDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LandmarkDetail {
  pageId: number;
  title: string;
  lat: number;
  lng: number;
  /** @nullable */
  extract?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  wikipediaUrl: string;
}